+++ app/javascript/maplibre/basemaps.js
@@ -148,6 +148,23 @@ export function basemaps () {
         glyphs: versatilesGlyphs
       }
     },
+    cyclosmTiles: {
+      style: {
+        version: 8,
+        sources: {
+          'raster-tiles': {
+            type: 'raster',
+            tiles: [
+              'https://b.tile-cyclosm.openstreetmap.fr/cyclosm/{z}/{x}/{y}.png'
+            ],
+            tileSize: 256,
+            attribution: '<a href="https://www.openstreetmap.org/copyright">© OpenStreetMap Contributors</a>'
+          }
+        },
+        layers: defaultRasterLayer,
+        glyphs: versatilesGlyphs
+      }
+    },
 
     // openfreemap.org
     openfreemapPositron: { style: 'https://tiles.openfreemap.org/styles/positron', font: 'Noto Sans Regular' },
@@ -159,6 +176,7 @@ export function basemaps () {
     versatilesColorful: { style: 'https://tiles.versatiles.org/assets/styles/colorful/style.json', font: 'noto_sans_regular' },
     versatilesGraybeard: { style: 'https://tiles.versatiles.org/assets/styles/graybeard/style.json', font: 'noto_sans_regular' },
     versatilesNeutrino: { style: 'https://tiles.versatiles.org/assets/styles/neutrino/style.json', font: 'noto_sans_regular' },
+    versatilesEclipse: { style: 'https://tiles.versatiles.org/assets/styles/eclipse/style.json', font: 'noto_sans_regular' },
 
     // Maptiler maps: https://docs.maptiler.com/sdk-js/api/map-styles/#mapstylelist
     // 3D Houses