+++ config/importmap.rb
@@ -40,6 +40,10 @@ pin "@maplibre/maplibre-gl-directions", preload: false, to: "@maplibre--maplibre
 # https://github.com/onthegomap/maplibre-contour
 pin "maplibre-contour", preload: false # @0.1.0
 
+# Elevation chart
+pin "chart.js", to: "chart-js.js", preload: false # @4.4.9
+pin "@kurkle/color", to: "@kurkle--color", preload: false # @0.3.4
+
 # render markdown
 pin "marked"# @15.0.8